Create supersized divs when called
The original supersized library generate 2 divs automatically:
 1. #supersized-loader
 2. #supersized

These 2 divs are generated even if the library is not called by
explicit `$('selector').supersized()` call. That is an unexpected
behaviour.

This commit movied the append statement into the supersized
plugin. The divs are now appended only if:

 1. $.supersized is called explicitly; and
 2. The #supersized div is previously not existed

The behaviour is a slightly different from before, but all other
behaviour reserved.